chore: merge state and env
merge bootstrap files
