+++ packages/rum-core/src/state.js
@@ -26,3 +26,5 @@
 const __DEV__ = process.env.NODE_ENV !== 'production'
 
 export { __DEV__ }
+
+export let state = {}
